requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
redis==5.0.1
redis==5.0.1
torch==2.2.0
transformers==4.35.0
scikit-learn==1.3.0
numpy==1.26.4
pandas==2.0.3
nltk==3.8.1
spacy==3.7.2
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
bleach==6.1.0
slowapi==0.1.9 